Refactor c_lapack implementation of linsolve interface
