+++ oif_impl/impl/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(qeq)
+add_subdirectory(linsolve)
