+++ oif_impl/impl/linsolve/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(c_lapack)
