+++ oif_impl/impl/linsolve/c_lapack/c_lapack.conf
@@ -1,0 +1,2 @@
+c
+liboif_linsolve_c_lapack.so
